fix(ci): python-coverage-shards — split oversized provisioner/retry modules under 1500-line gate
Recent PR fixes pushed provisioner.py and workspaces_retry.py over the first-party line-limit maintainability check; extract launch-cleanup mixin and retry payload helpers without behavior change.
