+++ src/awf/node/provisioner.py
@@ -16,13 +16,11 @@ re-raised so the caller can log/alert.
 
 from __future__ import annotations
 
-import asyncio
 from collections.abc import Awaitable, Callable, Mapping
 from datetime import UTC, datetime
 from pathlib import Path
 from typing import Any, Final, cast
 
-from sqlalchemy import func, select
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 
 from awf.adapters.model_selection import selected_runtime_model_for_defaults
@@ -44,15 +42,10 @@ from awf.db.enums import (
     WorkspaceStatus,
     parse_agent_runtime,
 )
-from awf.db.models import Workspace, WorkspaceEvent
+from awf.db.models import Workspace
 from awf.db.repositories import WorkspaceRepository
 from awf.db.repositories.base import (
     PRE_LAUNCH_FAILURE_EVENT_TYPE,
-    PROVISIONING_LAUNCHING_EVENT_TYPE,
-    PROVISIONING_LAUNCHING_REASON_CODE,
-    TERMINAL_RUNTIME_RELEASE_REVOKED_EVENT_TYPE,
-    TERMINAL_RUNTIME_RELEASE_REVOKED_REASON_CODE,
-    has_terminal_runtime_released_event,
 )
 from awf.node import provisioner_config as _provisioner_config
 from awf.node import provisioner_helpers as _provisioner_helpers
@@ -71,6 +64,7 @@ from awf.node.egress_policy import LocalEgressPlan, LocalEgressPolicyError, loca
 from awf.node.git_manager import GitManager, GitOperationError
 from awf.node.provisioner_cursor_preflight import ProvisionerCursorPreflightMixin
 from awf.node.provisioner_host_ports_check import ProvisionerHostPortCheckMixin
+from awf.node.provisioner_launch_cleanup import ProvisionerLaunchCleanupMixin
 from awf.node.provisioner_short_txn_helpers import ProvisionerShortTxnHelpersMixin
 from awf.node.stack_launcher import WorkspaceStackLauncher, WorkspaceStackLaunchRequest
 from awf.profiles.compose import profile_services
@@ -79,7 +73,6 @@ from awf.profiles.resolver import (
     ProfileResolutionError,
     resolve_workspace_profile,
 )
-from awf.service.controls_helpers import stop_project_containers
 from awf.service.secret_leases import (
     PROVISIONING_FAILED_REVOKE_REASON,
     SecretLeaseService,
@@ -116,12 +109,6 @@ _sync_feature_pr_pull_head_ref = _provisioner_helpers._sync_feature_pr_pull_head
 ProvisionerConfig = _provisioner_config.ProvisionerConfig
 ServiceStartupDiagnosticsCapturer = _provisioner_config.ServiceStartupDiagnosticsCapturer
 
-_MAX_REVOKE_EVENTS: Final = 3
-"""Maximum lifetime-total revoke events before recording an operator escalation event."""
-
-_ORPHAN_STOP_TIMEOUT_SECONDS: Final = 30.0
-"""Maximum time to spend stopping orphan containers after launch races cleanup."""
-
 _EXECUTION_CLAIM_FENCED_REASON_CODE: Final = "EXECUTION_CLAIM_FENCED"
 """Reason code logged when a stale provisioner is fenced by the execution-claim epoch."""
 
@@ -149,6 +136,7 @@ _parse_agent_runtime = parse_agent_runtime
 class Provisioner(
     ProvisionerCursorPreflightMixin,
     ProvisionerHostPortCheckMixin,
+    ProvisionerLaunchCleanupMixin,
     ProvisionerShortTxnHelpersMixin,
 ):
     """Orchestrate one workspace at a time, safely across concurrent provisions."""
@@ -1134,30 +1122,6 @@ class Provisioner(
                 ),
             )
 
-    async def _launch_lost_to_terminal_cleanup_best_effort(
-        self,
-        workspace_id: str,
-        *,
-        failure_context: str,
-    ) -> bool:
-        """Run the launch-cleanup race check without masking failure handling.
-
-        This wrapper is only for exception handlers. The normal post-launch
-        success path should keep calling `_launch_lost_to_terminal_cleanup`
-        directly so an indeterminate cleanup check cannot incorrectly proceed
-        to `ready`.
-        """
-        try:
-            return await self._launch_lost_to_terminal_cleanup(workspace_id)
-        except Exception:
-            _log.exception(
-                "provisioner.launch_lost_to_terminal_cleanup_check_failed",
-                workspace_id=workspace_id,
-                failure_context=failure_context,
-                reason_code="TERMINAL_CLEANUP_CHECK_FAILED",
-            )
-            return False
-
     async def _reject_unsupported_agent_runtime(
         self,
         *,
@@ -1329,188 +1293,6 @@ class Provisioner(
                 await session.commit()
         except Exception:  # pragma: no cover - defensive
             _log.exception("provisioner.mark_failed_failed", workspace_id=workspace_id)
-
-    async def _recheck_before_launch(self, workspace_id: str) -> bool:
-        """Recheck workspace status with a row lock and record a launch guard.
-
-        Unlike :meth:`_recheck_status`, this method holds a ``SELECT FOR UPDATE``
-        row lock while checking status and recording a
-        ``workspace.provisioning_launching`` event in the same transaction.  The
-        row lock serializes with concurrent cancel/stop/destroy operations that
-        also read the workspace row before transitioning it to a terminal state.
-
-        The current cancel/stop control operations use a *synchronous* project
-        stopper that blocks until containers are fully down, which guarantees the
-        stack is no longer consuming host ports before the ``terminal_runtime_released``
-        event is committed.  The ``provisioning_launching`` event serves as an
-        audit trail marker but is not currently read by cancel/stop logic; a
-        future async stopper would need to check this event before emitting
-        ``terminal_runtime_released`` to avoid a race with a launch that has
-        already committed.
-
-        Returns ``True`` when the workspace is still ``provisioning`` and the
-        launch-guard event was recorded; ``False`` otherwise.
-        """
-        async with self._session_factory() as session:
-            repo = WorkspaceRepository(session)
-            ws = await repo.get_for_update(workspace_id)
-            if ws is None:
-                _log.warning(
-                    "provisioner.skip_unknown_before_launch",
-                    workspace_id=workspace_id,
-                )
-                await session.commit()
-                return False
-            if ws.status != WorkspaceStatus.provisioning.value:
-                await self._record_stale_action_skip(
-                    repo,
-                    ws,
-                    action="provision",
-                    expected=WorkspaceStatus.provisioning,
-                    reason_code="PROVISIONER_STALE_STATUS",
-                )
-                await session.commit()
-                return False
-            await repo.add_event(
-                ws,
-                event_type=PROVISIONING_LAUNCHING_EVENT_TYPE,
-                reason_code=PROVISIONING_LAUNCHING_REASON_CODE,
-                payload={"workspace_id": workspace_id},
-            )
-            await session.commit()
-            return True
-
-    async def _launch_lost_to_terminal_cleanup(self, workspace_id: str) -> bool:
-        """Check whether terminal cleanup won while the stack was launching.
-
-        When an operator force-destroys the workspace after
-        ``_recheck_before_launch`` commits its ``provisioning_launching``
-        guard but before ``_stack_launcher.launch`` actually starts,
-        ``destroy_workspace(force=True)`` can see the pre-published
-        ``compose_project_name``, run cleanup before any containers exist,
-        transition to ``destroyed``, and record
-        ``workspace.terminal_runtime_released``.  The provisioner then
-        still launches the stack, leaving running containers that future
-        host-port admission ignores because the release event exists.
-
-        This method detects that outcome: if ``terminal_runtime_released``
-        was recorded while we were launching, stop the just-launched
-        containers and return ``True`` so the caller aborts without
-        transitioning to ``ready``.
-
-        The DB session is released before Docker I/O and reacquired
-        afterwards so that a slow or unresponsive Docker daemon does not
-        hold a pool connection.  The row lock (``get_for_update``) is
-        acquired only for the brief ``add_event`` / ``commit`` step
-        after Docker I/O completes.
-
-        Mitigations for pool exhaustion: (1) the DB session is released
-        before Docker I/O; (2) ``stop_project_containers`` is bounded by
-        ``_ORPHAN_STOP_TIMEOUT_SECONDS``; (3) the pool ``max_size`` should
-        account for at most one concurrent orphan-stop per node.
-
-        Returns ``True`` when terminal cleanup won and containers were
-        stopped; ``False`` when the workspace is still clear to proceed.
-        """
-        compose_project = f"awf_{workspace_id}"
-        prior_status: str | None = None
-        async with self._session_factory() as session:
-            repo = WorkspaceRepository(session)
-            ws = await repo.get(workspace_id)
-            if ws is None:
-                return False
-            released = await has_terminal_runtime_released_event(session, workspace_id)
-            if not released:
-                return False
-            prior_status = ws.status
-
-        _log.warning(
-            "provisioner.launch_lost_to_terminal_cleanup",
-            workspace_id=workspace_id,
-            reason_code="TERMINAL_CLEANUP_WON_DURING_LAUNCH",
-        )
-        orphan_stopped = True
-        orphan_stop_error: str | None = None
-        try:
-            await asyncio.wait_for(
-                stop_project_containers(compose_project),
-                timeout=_ORPHAN_STOP_TIMEOUT_SECONDS,
-            )
-        except TimeoutError:
-            orphan_stopped = False
-            orphan_stop_error = (
-                f"stop_project_containers timed out after {_ORPHAN_STOP_TIMEOUT_SECONDS:g}s"
-            )
-            _log.warning(
-                "provisioner.orphan_container_stop_timeout",
-                workspace_id=workspace_id,
-                reason_code="ORPHAN_STOP_TIMEOUT",
-                timeout_seconds=_ORPHAN_STOP_TIMEOUT_SECONDS,
-            )
-        except Exception as exc:
-            orphan_stopped = False
-            orphan_stop_error = str(exc)
-            _log.exception(
-                "provisioner.orphan_container_stop_failed",
-                workspace_id=workspace_id,
-                reason_code="ORPHAN_STOP_FAILED",
-            )
-        async with self._session_factory() as session:
-            repo = WorkspaceRepository(session)
-            ws = await repo.get_for_update(workspace_id)
-            if ws is None:
-                return True
-            payload: dict[str, object] = {
-                "action": "provision",
-                "expected_status": WorkspaceStatus.provisioning.value,
-                "actual_status": prior_status,
-                "orphan_containers_stopped": orphan_stopped,
-            }
-            if orphan_stop_error is not None:
-                payload["orphan_stop_error"] = orphan_stop_error
-                revoke_count_result = await session.execute(
-                    select(func.count()).where(
-                        WorkspaceEvent.workspace_id == workspace_id,
-                        WorkspaceEvent.event_type == TERMINAL_RUNTIME_RELEASE_REVOKED_EVENT_TYPE,
-                        WorkspaceEvent.reason_code == TERMINAL_RUNTIME_RELEASE_REVOKED_REASON_CODE,
-                    )
-                )
-                revoke_count = revoke_count_result.scalar() or 0
-                await repo.add_event(
-                    ws,
-                    event_type=TERMINAL_RUNTIME_RELEASE_REVOKED_EVENT_TYPE,
-                    reason_code=TERMINAL_RUNTIME_RELEASE_REVOKED_REASON_CODE,
-                    payload={
-                        "workspace_id": workspace_id,
-                        "orphan_stop_error": orphan_stop_error,
-                    },
-                )
-                if revoke_count + 1 >= _MAX_REVOKE_EVENTS:
-                    await repo.add_event(
-                        ws,
-                        event_type="workspace.stale_action_skipped",
-                        reason_code="REVOKE_CAP_REACHED",
-                        payload={
-                            "workspace_id": workspace_id,
-                            "revoke_count": revoke_count + 1,
-                            "orphan_stop_error": orphan_stop_error,
-                            "message": (
-                                f"{revoke_count + 1} lifetime-total revoke events; "
-                                "operator intervention may be required to stop "
-                                "orphan containers and free host ports. "
-                                "Revoke events will continue to be recorded "
-                                "until the runtime is released."
-                            ),
-                        },
-                    )
-            await repo.add_event(
-                ws,
-                event_type="workspace.stale_action_skipped",
-                reason_code="TERMINAL_CLEANUP_WON_DURING_LAUNCH",
-                payload=payload,
-            )
-            await session.commit()
-        return True
 
     async def _record_stale_action_skip(
         self,
